Player Update and Bug fixes

+++ src/Modules/RenderModule.java
@@ -17,7 +17,7 @@ public class RenderModule {
         g.drawString(text, x, y);
     }
 
-    public static float clamp(float f, float min, float max){
+    public static float clamp(float f, int min, int max){
         if(f > max){
             return max;
         } else if(f < min){
